api/requirements.txt
fastapi==0.68.1
colorama==0.4.4
pymongo==3.12.0
uvicorn==0.15.0
argon2-cffi==21.1.0